refactor(domain): move deriveAccentVariant to colorTheme domain
