+++ src/domain/colorTheme/index.ts
@@ -193,6 +193,17 @@ export const PREDEFINED_PALETTES: ColorPalette[] = [
   },
 ]
 
+export function deriveAccentVariant(hex: string, mode: 'dark' | 'light'): string {
+  const r = parseInt(hex.slice(1, 3), 16)
+  const g = parseInt(hex.slice(3, 5), 16)
+  const b = parseInt(hex.slice(5, 7), 16)
+  const mix = mode === 'dark' ? 180 : 80
+  const nr = Math.round(r + (mix - r) * 0.45)
+  const ng = Math.round(g + (mix - g) * 0.45)
+  const nb = Math.round(b + (mix - b) * 0.45)
+  return `#${nr.toString(16).padStart(2, '0')}${ng.toString(16).padStart(2, '0')}${nb.toString(16).padStart(2, '0')}`
+}
+
 export function getPaletteById(
   id: string,
   customs: CustomPalette[] = [],
